[*] split exception.hpp to exception.cpp
