+++ src/co/exception.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <exception>
-#include <iostream>
+#include <iosfwd>
 #include <co/status_codes.hpp>
 
 namespace co
@@ -17,14 +17,7 @@ public:
         : co::status_code(co::other)
     {}
 
-    error_desc(const co::status_code& status, const char* desc)
-        : co::status_code(status)
-        , _desc(desc)
-    {}
-
-    error_desc(const co::status_code& status)
-        : co::status_code(status)
-    {}
+    error_desc(const co::status_code& status, const char* desc = "");
 
     error_desc(const co::exception& coexc);
 
@@ -43,6 +36,17 @@ private:
     const char* _desc = "";  // error description, MUST HAVE static lifetime duration
 };
 
+/// \brief co_lib exception class that holds co::status_code and custom description as not owned const char*
+///
+/// NOTE: const char* desc must have static lifetime
+/// Usage:
+/// \code
+///     throw co::exception(co::other, "something bad happened");
+/// \endcode
+/// How to choose between co::result and co::exception? General rule is the next: use return code (co::result) for
+/// expected outcomes of the operation: network errors, operation cancellation, channel is full. And use co::exception
+/// (or more general std::exception) as a not expected error. like class misuse, unexpected errors of underlying
+/// systems, allocation errors, logic errors, etc.
 class exception : public std::exception
 {
 public:
@@ -73,15 +77,6 @@ private:
     error_desc _edesc;
 };
 
-inline std::ostream& operator<<(std::ostream& out, const co::exception& coexc)
-{
-    out << coexc.status().category_name() << "::" << coexc.status().message() << "(" << coexc.status().message() << ") "
-        << coexc.what();
-    return out;
-}
-
-inline error_desc::error_desc(const co::exception& coexc)
-    : error_desc(coexc.status(), coexc.what())
-{}
+std::ostream& operator<<(std::ostream& out, const co::exception& coexc);
 
 }  // namespace co
