+++ src/co/thread.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <iostream>
 #include <co/func.hpp>
 #include <co/std.hpp>
 #include <co/until.hpp>
